418 I'm a teapot
Jenkins hashes of some renamed natives have been removed to avoid confusion and not break tools that validate names via an existing jhash.

This can always be changed back later if necessary, potentially via a new field, even (to indicate a name was changed, etc). But for now this should do.

Who would've thought a teapot would be the one to breach Rockstar's security and give us this amazing stuff after almost a decade?